requirements.txt
-i https://pypi.python.org/simple
astroid==1.6.1
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
firebase-admin==2.9.0
future==0.16.0
google-api-core==1.0.0
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.8.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.10.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
protobuf==3.5.2
pyasn1-modules==0.2.1
pyasn1==0.4.2
pylint==1.8.2
pymongo==3.6.1
python-telegram-bot==10.0.1
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
wrapt==1.10.11
